[hello] fix start routines and stop cowboy listener routines
Summary:
Someday, after update all applications to yang-based hello (that is hello2 at the moment)
we will not differentiate between hello and hello2. This split bring only problem (maintaining 2 version and so on).

There is a branch yang, which is rebased to the latests hello2/master, renamed back to hello.
This differential, fix the listener problem, where child listener is started throw ranch supervisor and added to hello_binding_sup as child.
If you try to stop hello with started http listener, it will hangs infinity, because application master tries to stop supervisor, that tries
to stop listener childs, but this listener childs doesn't have hello_binding supervisor as parent, and therefor ignores it as shutdown from
hello_binding doesn't match with a parent id (which is a ranch supervisor).

Fixing it, throw adding new callback, and differentiate behaviour of listener_specification, at it can start it in own supervisor or
add to hello_binding supervisor(as for example zeromq, which doesn't have own connection supervisor).

Test Plan:
tetrapak test

TEST COMPLETE, 65 ok, 1 failed, 6 skipped of 72 test cases
6 skipped is ok.
1 failed throw a bug in yang.

Reviewers: aschultz, vtarasen, orixmann, skattelmann, akuleshov

Reviewed By: skattelmann

CC: mkraemer, aschultz

Differential Revision: https://review.tpip.net/D1270